+++ NEWS.md
@@ -1,5 +1,11 @@
 # forestmangr development
 
+# forestmangr 0.9.2
+
+* updated tree_summarise to sum volume with and without bark using 2 new arguments
+
+* fixed vol_summarise handlying of optional variables
+
 # forestmangr 0.9.1
 
 * Graybill_F is now graybill_f;